Rename params->configs
