+++ GPTifier/src/api.cpp
@@ -1,5 +1,5 @@
 #include "api.hpp"
-#include "params.hpp"
+#include "configs.hpp"
 
 #include <stdexcept>
