+++ GPTifier/src/configs.cpp
@@ -1,4 +1,4 @@
-#include "params.hpp"
+#include "configs.hpp"
 #include "utils.hpp"
 
 #include <filesystem>
